♻️ refactor: Event 도메인 디렉토리 정리
- Event 도메인 디렉토리 컨벤션에 맞게 정리
- Repository 레이어 추가
Issue Resolved: #40
