+++ back/src/event/controller/event.controller.ts
@@ -1,5 +1,5 @@
 import { Controller } from '@nestjs/common';
-import { EventService } from './event.service';
+import { EventService } from '../service/event.service';
 
 @Controller('event')
 export class EventController {
