+++ back/src/event/event.module.ts
@@ -1,9 +1,14 @@
 import { Module } from '@nestjs/common';
-import { EventService } from './event.service';
-import { EventController } from './event.controller';
+import { EventService } from './service/event.service';
+import { EventController } from './controller/event.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { EventRepository } from './repository/event.reposiotry';
 
 @Module({
+  imports: [
+    TypeOrmModule.forFeature([Event]),
+  ],
   controllers: [EventController],
-  providers: [EventService],
+  providers: [EventService, EventRepository],
 })
 export class EventModule {}
